restructure some components to functional, removes link component
