+++ clients/fides-js/src/fides.ts
@@ -119,10 +119,9 @@ if (typeof window !== "undefined") {
 
 // Export everything from ./lib/* to use when importing fides.mjs as a module
 export * from "./lib/consent";
-export * from "./lib/consent-banner";
+export * from "./components/ConsentBanner";
 export * from "./lib/consent-config";
 export * from "./lib/consent-context";
-export * from "./lib/consent-link";
 export * from "./lib/consent-types";
 export * from "./lib/consent-utils";
 export * from "./lib/consent-value";
