+++ clients/privacy-center/public/fides-js-components-demo.html
@@ -37,7 +37,7 @@
       <pre id="consent-json"></pre>
 
       <h2>Consent Banner Link</h2>
-      <pre id="fides-consent-link"></pre>
+      <a href="#" id="fides-consent-link">Manage Consent</a>
 
       <h2>Consent Banner Options</h2>
       <pre id="consent-banner-options"></pre>
