+++ clients/privacy-center/public/fides-js-demo.html
@@ -31,7 +31,7 @@
       <pre id="consent-json"></pre>
 
       <h2>Consent Banner Link</h2>
-      <pre id="fides-consent-link"></pre>
+      <a href="#" id="fides-consent-link">Manage Consent</a>
 
       <h2>Consent Banner Options</h2>
       <pre id="consent-banner-options"></pre>
